(a) moved NDimBoundingBox from utils.py to romc.py (b) added a test for RomcPosterior (c) visualize_region() from RomcPosterior to ROMC class
